ghettowm: refactor, create HWND_MESSAGE window for shell msgs
We're now creating a "fake" window for the sake of capturing
shell messages by registering ourself as shell hook window.

This gives us access to WM_SHELLHOOK messages, which we'll use in the
near future to know when windows are being created by other
applications.

We'll still need an initial EnumWindows to begin any layout/tiling.