ghetto(main): minor progress on managing individual windows
There's a lot more to do here, I toyed around with resizing &
positioning of windows but haven't actually written anything useful
in that regard.

ghetto is now tracking windows per desktop on initial run,
and just needs to start handling them from the shellhook.

I may not get around to implementing tiling for a while..
As it is, being able to switch between them is probably
a good enough boon for me at the moment.

What I'd like to do in the future is implement a tray icon, allowing an
easy quit/restart when conf changes. I think that a hot-restart may
be problematic given the way it runs now, but we'll see.